Add profile service (#4)
* feat: add profile service

* feat: refactor

* feat: fix validations

* feat: add validation for user fields

+++ src/main/java/tech/ayot/ticket/backend/repository/user/UserProductRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import tech.ayot.ticket.backend.model.user.UserProduct;
 
 @Repository
-public interface UserProductRepository extends JpaRepository<UserProduct, Integer> {
+public interface UserProductRepository extends JpaRepository<UserProduct, Long> {
 
     UserProduct findUserProductByUserUsername(String username);
 }
